pyproject.toml
[tool.poetry]
name = "NoteBook"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
fastapi-jwt-auth = "^0.5.0"
fastapi-mail = "^1.4.1"
bcrypt = "4.0.1"
slowapi = "^0.1.9"
cloudinary = "^1.40.0"
sphinx = "^7.3.7"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
